\name{pbdADIOS-package}
\alias{pbdADIOS-package}
\alias{pbdADIOS}
\docType{package}
\title{Programming with Big Data -- Interface to ADIOS}
\description{
  pbdADIOS is a package to connect R with ADIOS parallel middleware
  for IO.
}
\details{
\tabular{ll}{
  Package: \tab pbdADIOS\cr
  Type: \tab Package\cr
  License: \tab Mozilla Public License 2.0\cr
  LazyLoad: \tab yes\cr
}

  This package requires a build of the ADIOS library with
  
  \code{MPICC=mpicc -fPIC}
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\keyword{package}
