% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.advance.step}
\alias{adios.advance.step}
\title{Advance the current step of a stream.}
\usage{
adios.advance.step(adios.file.ptr, adios.last, adios.timeout.sec)
}
\arguments{
\item{adios.timeout.sec}{}
}
\description{
For files opened as file, stepping has no effect.
}

