% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{adios.declare.group}
\alias{adios.declare.group}
\title{Declare a ADIOS group.}
\usage{
adios.declare.group(adios.groupname, adios.timeindex, adios.flag)
}
\arguments{
\item{adios.flag}{}
}
\value{
adios group id pointer
}

