% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{adios.define.var}
\alias{adios.define.var}
\title{Define a ADIOS variable}
\usage{
adios.define.var(adios.group_p, adios.varname, adios.path, adios.type,
  adios.localdim, adios.globaldim, adios.localoffset)
}
\arguments{
\item{adios.localoffset}{}
}
\value{
a variable ID, which can be used in adios_write_byid()
   0 return value indicates an error
}

