% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{adios.finalize}
\alias{adios.finalize}
\title{adios.finalize}
\usage{
adios.finalize(comm = pbdMPI::.SPMD.CT$comm)
}
\arguments{
\item{comm}{The MPI communicator.}
}
\value{
\code{NULL}
}
\description{
Shut down ADIOS.
}

