% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{adios.group.size}
\alias{adios.group.size}
\title{This function passes the size of the group to the internal ADIOS transport structure 
to facilitate the internal buffer management and to construct the group index table}
\usage{
adios.group.size(adios.file_p, adios.groupsize)
}
\arguments{
\item{adios.groupsize}{}
}
\value{
R_adios_total_size
}
\description{
This function passes the size of the group to the internal ADIOS transport structure 
to facilitate the internal buffer management and to construct the group index table
}

