% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adios.init.noxml}
\alias{adios.init.noxml}
\title{Write Initialization}
\usage{
adios.init.noxml(comm = pbdMPI::.SPMD.CT$comm)
}
\arguments{
\item{adios.init.noxml}{A communicator number.
Description of the variable.}
}
\value{
Description of the return.
}
\description{
A description of what the function does.
}
\details{
A more detailed description, if you like.
}
\examples{
\dontrun{
library(pbdADIOS)
whatever <- adios.init.noxml(stuff)
}
}
\seealso{
\code{\link{adios.open}}
}

