% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.inq.var.blockinfo}
\alias{adios.inq.var.blockinfo}
\title{Get the block-decomposition of the variable about how it is stored in 
the file or stream.}
\usage{
adios.inq.var.blockinfo(adios.file.ptr, adios.varinfo)
}
\arguments{
\item{adios.varinfo}{}
}

