% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.inq.var.byid}
\alias{adios.inq.var.byid}
\title{Inquiry a variable by index.}
\usage{
adios.inq.var.byid(adios.file.ptr, adios.varid)
}
\arguments{
\item{adios.varid}{}
}
\value{
adios_var_info pointer
}

