% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.inq.var}
\alias{adios.inq.var}
\title{Inquiry about a variable.}
\usage{
adios.inq.var(adios.file.ptr, adios.varname)
}
\arguments{
\item{adios.varname}{}
}
\value{
adios_var_info pointer
}
\description{
This function does not read anything from the file but processes info
already in memory after fopen.
}

