% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{adios.open}
\alias{adios.open}
\title{Open or to append to an output file.}
\usage{
adios.open(adios.groupname, adios.filename, adios.mode,
  comm = pbdMPI:::.pbd_env$SPMD.CT$comm)
}
\arguments{
\item{comm}{}
}
\value{
adios file id pointer
}
\description{
modes = "r" = "read", "w" = "write", "a" = "append", "u" = "update"
}

