% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.read.init.method}
\alias{adios.advance.step}
\alias{adios.errno}
\alias{adios.free.varinfo}
\alias{adios.inq.var}
\alias{adios.inq.var.blockinfo}
\alias{adios.inq.var.byid}
\alias{adios.perform.reads}
\alias{adios.read.close}
\alias{adios.read.finalize.method}
\alias{adios.read.init.method}
\alias{adios.read.open}
\alias{adios.release.step}
\alias{adios.schedule.read}
\alias{adios.selection.boundingbox}
\alias{custom.data.access}
\alias{custom.inq.var.dims}
\alias{custom.inq.var.ndim}
\title{R wrapper of ADIOS read API}
\usage{
adios.read.init.method(adios.read.method, comm = .pbd_env$SPMD.CT$comm,
  params)

adios.read.open(adios.filename, adios.read.method,
  comm = .pbd_env$SPMD.CT$comm, adios.lockmode, adios.timeout.sec)

adios.inq.var(adios.file.ptr, adios.varname)

adios.inq.var.byid(adios.file.ptr, adios.varid)

adios.free.varinfo(adios.varinfo)

adios.inq.var.blockinfo(adios.file.ptr, adios.varinfo)

custom.inq.var.ndim(adios.varinfo)

custom.inq.var.dims(adios.varinfo)

adios.selection.boundingbox(adios.ndim, adios.start, adios.count)

adios.schedule.read(adios.varinfo, adios.start, adios.count, adios.file.ptr,
  adios.selection, adios.varname, adios.from.steps, adios.nsteps)

adios.perform.reads(adios.file.ptr, adios.blocking)

custom.data.access(adios.data, adios.selection, adios.varinfo)

adios.release.step(adios.file.ptr)

adios.advance.step(adios.file.ptr, adios.last, adios.timeout.sec)

adios.read.close(adios.file.ptr)

adios.read.finalize.method(adios.read.method)

adios.errno()
}
\description{
For more details of these APIs, please refer to https://github.com/ornladios/ADIOS/blob/master/src/public/adios_read_v2.h
}

