% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.read.close}
\alias{adios.read.close}
\title{Close an adios file.}
\usage{
adios.read.close(adios.file.ptr)
}
\arguments{
\item{adios.file.ptr}{}
}
\description{
It will free the content of the underlying data structures and the fp pointer itself.
}

