% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.read.finalize.method}
\alias{adios.read.finalize.method}
\title{Finalize the selected method.}
\usage{
adios.read.finalize.method(adios.read.method)
}
\arguments{
\item{adios.read.method}{}
}
\description{
Required for all methods that are initialized.
}

