% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adios_read_str.r
\name{adios.read.init.method}
\alias{adios.read.init.method}
\title{Read Initialization}
\usage{
adios.read.init.method(adios.read.method, comm = pbdMPI::.SPMD.CT$comm,
  params)
}
\arguments{
\item{adios.read.method}{Description of the variable.}

\item{comm}{A communicator number.}

\item{params}{Description of the variable.}
}
\value{
Description of the return.
}
\description{
A description of what the function does.
}
\details{
A more detailed description, if you like.
}
\examples{
\dontrun{
library(pbdADIOS)
whatever <- adios.read.init.method(stuff)
}
}
\seealso{
\code{\link{adios.read.open}}
}

