% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.read.open}
\alias{adios.read.open}
\title{Open an adios file/stream as a stream.}
\usage{
adios.read.open(adios.filename, adios.read.method,
  comm = .pbd_env$SPMD.CT$comm, adios.lockmode, adios.timeout.sec)
}
\arguments{
\item{adios.timeout.sec}{}
}
\value{
adios file pointer
}
\description{
Only one step at a time can be read. The list of variables will change when
advancing the step if the writing application writes different variables at
different times.
}

