% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adios_read.r
\name{adios.read}
\alias{adios.read}
\alias{adios.read.close}
\alias{adios.read.finalize.method}
\alias{adios.read.init.method}
\alias{adios.read.open}
\title{Read Initialization}
\usage{
adios.read.init.method(adios.read.method, comm = pbdMPI::.SPMD.CT$comm,
  params)

adios.read.open(adios.filename, adios.read.method,
  comm = pbdMPI::.SPMD.CT$comm, adios.lockmode, adios.timeout.sec)

adios.read.close(adios.file.ptr)

adios.read.finalize.method(adios.read.method)
}
\arguments{
\item{adios.read.method}{TODO}

\item{comm}{TODO}

\item{params}{TODO}

\item{adios.filename}{TODO}

\item{adios.lockmode}{TODO}

\item{adios.timeout.sec}{TODO}

\item{adios.file.ptr}{TODO}
}
\description{
TODO a quick description
}
\details{
TODO more details
}
\examples{
\dontrun{
TODO
}

}

