% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.schedule.read}
\alias{adios.schedule.read}
\title{Schedule reading a variable (slice) from the file.}
\usage{
adios.schedule.read(adios.varinfo, adios.start, adios.count, adios.file.ptr,
  adios.selection, adios.varname, adios.from.steps, adios.nsteps)
}
\arguments{
\item{adios.nsteps}{}
}
\value{
pointer to the memory to hold dsata of the variable.
}
\description{
You need to call adios.perform.reads to do the reading.
}

