% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{adios.select.method}
\alias{adios.select.method}
\title{Select a I/O method for a ADIOS group.}
\usage{
adios.select.method(adios.group_p, adios.method, adios.params, adios.basepath)
}
\arguments{
\item{adios.basepath}{}
}

