% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.selection.boundingbox}
\alias{adios.selection.boundingbox}
\title{adios_selection_bounding_box API.}
\usage{
adios.selection.boundingbox(adios.ndim, adios.start, adios.count)
}
\arguments{
\item{adios.count}{}
}
\value{
adios_selection pionter
}

