% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{adios.init.noxml}
\alias{adios.close}
\alias{adios.declare.group}
\alias{adios.define.var}
\alias{adios.finalize}
\alias{adios.group.size}
\alias{adios.init.noxml}
\alias{adios.open}
\alias{adios.select.method}
\alias{adios.set.max.buffersize}
\alias{adios.write}
\title{R wrapper of ADIOS write API}
\usage{
adios.init.noxml(comm = .pbd_env$SPMD.CT$comm)

adios.set.max.buffersize(adios.max.buffersize)

adios.declare.group(adios.groupname, adios.timeindex, adios.flag)

adios.select.method(adios.group_p, adios.method, adios.params, adios.basepath)

adios.define.var(adios.group_p, adios.varname, adios.path, adios.type,
  adios.localdim, adios.globaldim, adios.localoffset)

adios.open(adios.groupname, adios.filename, adios.mode,
  comm = .pbd_env$SPMD.CT$comm)

adios.group.size(adios.file_p, adios.groupsize)

adios.write(adios.file_p, adios.varname, adios.var)

adios.close(adios.file_p)

adios.finalize(comm = .pbd_env$SPMD.CT$comm)
}
\description{
For more details of these APIs, please refer to https://github.com/ornladios/ADIOS/blob/master/src/public/adios.h
}

