% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bp.attr}
\alias{bp.attr}
\title{bp.attr}
\usage{
bp.attr(adios.attrname, data)
}
\arguments{
\item{adios.attrname}{adios attribute name}

\item{data}{value of the attribute}
}
\description{
This function is used to define attributes for a bp file.
}
\examples{
today <- Sys.Date()
bp.attr("date", format(today, format="\%B \%d \%Y"))
bp.attr("README", "This is just a test.")

}

