% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R, R/bp_bak.R
\name{bp.create}
\alias{bp.create}
\title{create adios group and store file name}
\usage{
bp.create(adios.filename, adios.groupname = "bp.group", buffer.size = 20,
  comm = .pbd_env$SPMD.CT$comm)

bp.create(adios.filename, adios.groupname = "bp.group", buffer.size = 20,
  comm = .pbd_env$SPMD.CT$comm)
}
\arguments{
\item{comm}{}

\item{comm}{}
}
\value{
group pointer.

group pointer.
}

