% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.stage.R
\name{bp.stage.read}
\alias{bp.stage.read}
\title{staging read variables}
\usage{
bp.stage.read(adios.filename, varname, FUN = print,
  comm = .pbd_env$SPMD.CT$comm, p = comm.size(.pbd_env$SPMD.CT$comm),
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))
}
\arguments{
\item{adios.rank}{}
}
\value{
variable values. If start and count are not specified, all values will be returned.
}

