% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bp.var}
\alias{bp.var}
\title{bp.var}
\usage{
bp.var(adios.varname, data)
}
\arguments{
\item{adios.varname}{variable name}

\item{data}{value of the variable}
}
\description{
This function is used to register variable names and values.
}
