% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{custom.data.access}
\alias{custom.data.access}
\title{Copy the scheduled reads from C object to R object.}
\usage{
custom.data.access(adios.data, adios.selection, adios.varinfo)
}
\arguments{
\item{adios.varinfo}{}
}
\value{
the data read by adios
}

