% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pbdADIOS-package}
\alias{pbdADIOS-package}
\title{logisticPCA-package}
\description{
Package to read from ADIOS, including staging.
}
\author{
Pragneshkumar Patel, George Ostrouchov, Wei-Chen Chen, Drew Schmidt
}
\keyword{package}

