% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pbdADIOS.R
\docType{package}
\name{pbdADIOS}
\alias{pbdADIOS}
\alias{pbdADIOS-package}
\title{pbdADIOS: an R wrapper for ADIOS}
\description{
Utilities to read from ADIOS, including staging.
}
\author{
The pbdR Core Team
}
\keyword{pbdADIOS}
