% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.read.init.method}
\alias{adios.advance.step}
\alias{adios.errno}
\alias{adios.free.varinfo}
\alias{adios.inq.var}
\alias{adios.inq.var.blockinfo}
\alias{adios.perform.reads}
\alias{adios.read.close}
\alias{adios.read.finalize.method}
\alias{adios.read.init.method}
\alias{adios.read.open}
\alias{adios.schedule.read}
\alias{adios.selection.boundingbox}
\alias{custom.data.access}
\alias{custom.inq.var.dims}
\alias{custom.inq.var.ndim}
\title{read init}
\usage{
adios.read.init.method(adios.read.method, comm = .pbd_env$SPMD.CT$comm,
  params)

adios.read.open(adios.filename, adios.read.method,
  comm = .pbd_env$SPMD.CT$comm, adios.lockmode, adios.timeout.sec)

adios.inq.var(adios.file.ptr, adios.varname)

adios.free.varinfo(adios.varinfo)

adios.inq.var.blockinfo(adios.file.ptr, adios.varinfo)

custom.inq.var.ndim(adios.varinfo)

custom.inq.var.dims(adios.varinfo)

adios.selection.boundingbox(adios.ndim, adios.start, adios.count)

adios.schedule.read(adios.varinfo, adios.start, adios.count, adios.file.ptr,
  adios.selection, adios.varname, adios.from.steps, adios.nsteps)

adios.perform.reads(adios.file.ptr, adios.blocking)

custom.data.access(adios.data, adios.selection, adios.varinfo)

adios.advance.step(adios.file.ptr, adios.last, adios.timeout.sec)

adios.read.close(adios.file.ptr)

adios.read.finalize.method(adios.read.method)

adios.errno()
}
\arguments{
\item{adios.read.method}{}

\item{params}{}

\item{adios.timeout.sec}{}

\item{adios.file.ptr}{}

\item{adios.varname}{}

\item{adios.varinfo}{}

\item{adios.count}{}

\item{adios.nsteps}{}

\item{adios.blocking}{}

\item{adios.varinfo}{}

\item{adios.varinfo}{}

\item{adios.varinfo}{}

\item{adios.varinfo}{}

\item{adios.timeout.sec}{}
}
\value{
adios file pointer

adios_var_info pointer

number of dimensions

size of each dimensions

adios_selection pionter

pointer to the memory to hold dsata of the variable.

the data read by adios

error number
}
\description{
Initialize a reading method before opening a file/stream with using the method.

Only one step at a time can be read. The list of variables will change when
advancing the step if the writing application writes different variables at
different times.

This function does not read anything from the file but processes info
already in memory after fopen.

You need to call adios.perform.reads to do the reading.

For files opened as file, stepping has no effect.

It will free the content of the underlying data structures and the fp pointer itself.

Required for all methods that are initialized.
}

