% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{adios.attr.read}
\alias{adios.attr.read}
\title{print all adios attributes}
\usage{
adios.attr.read(adios.rank = comm.rank(.pbd_env$SPMD.CT$comm), adios.fp)
}
\arguments{
\item{adios.rank}{comm rank}

\item{adios.fp}{}
}

