% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adios_inq_var.r
\name{adios.inq.var}
\alias{adios.inq.var}
\alias{adios.inq.var.blockinfo}
\alias{custom.inq.var.dims}
\alias{custom.inq.var.ndim}
\title{???}
\usage{
adios.inq.var(adios.file.ptr, adios.varname)

custom.inq.var.ndim(adios.varinfo)

custom.inq.var.dims(adios.varinfo)

adios.inq.var.blockinfo(adios.file.ptr, adios.varinfo)
}
\arguments{
\item{adios.file.ptr}{TODO}

\item{adios.varname}{TODO}

\item{adios.varinfo}{TODO}
}
\description{
TODO a quick description
}
\details{
TODO more details
}
\examples{
\dontrun{
TODO
}

}

