% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{adios.read.init.method}
\alias{adios.read.init.method}
\title{read init}
\usage{
adios.read.init.method(adios.read.method, comm = .pbd_env$SPMD.CT$comm,
  params)
}
\arguments{
\item{params}{}
}
\description{
Initialize a reading method before opening a file/stream with using the method.
}

