% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{adios.var.attr.read}
\alias{adios.var.attr.read}
\title{Print adios variables and attributes}
\usage{
adios.var.attr.read(adios.rank = comm.rank(.pbd_env$SPMD.CT$comm), adios.fp)
}
\arguments{
\item{adios.rank}{comm rank}

\item{adios.fp}{adios file pointer}
}
\description{
This function is used to print all adios variables with related attributes in a bp file.
}
