% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R, R/bp_bak.R
\name{bp.dump}
\alias{bp.dump}
\title{dump variables}
\usage{
bp.dump(adios.filename, comm = .pbd_env$SPMD.CT$comm,
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))

bp.dump(adios.filename, comm = .pbd_env$SPMD.CT$comm,
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))
}
\arguments{
\item{adios.rank}{}

\item{adios.rank}{}
}
\value{
a list of variables

a list of variables
}

