% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bp.read}
\alias{bp.read}
\title{read variable}
\usage{
bp.read(adios.filename, varname, start = NULL, count = NULL,
  comm = .pbd_env$SPMD.CT$comm, p = comm.size(.pbd_env$SPMD.CT$comm),
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))
}
\arguments{
\item{adios.rank}{}
}
\value{
variable values. If start and count are not specified, all values will be returned.
}

