% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.stage.R
\name{bp.stage.read}
\alias{bp.stage.read}
\title{bp.stage.read}
\usage{
bp.stage.read(adios.filename, varname, FUN = print,
  comm = .pbd_env$SPMD.CT$comm, p = comm.size(.pbd_env$SPMD.CT$comm),
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))
}
\arguments{
\item{adios.filename}{adios filename}

\item{varname}{variable name}

\item{comm}{}

\item{adios.rank}{}
}
\value{
variable values. If start and count are not specified, all values will be returned.
}
\description{
This function is used to read the variables using staging method. 
Current staging method used is FLEXPATH. The user can provide an function to handle
the data read from writing stream. If the function is not provided, the data will be
printed by default.
}

