% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bp.write}
\alias{bp.write}
\title{write variables to bp file}
\usage{
bp.write(comm = .pbd_env$SPMD.CT$comm, p = comm.size(.pbd_env$SPMD.CT$comm),
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))
}
\arguments{
\item{adios.rank}{}
}

