% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R, R/bp_bak.R
\name{bpls}
\alias{bpls}
\title{print variables and attributes}
\usage{
bpls(adios.filename, comm = .pbd_env$SPMD.CT$comm,
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))

bpls(adios.filename, comm = .pbd_env$SPMD.CT$comm,
  adios.rank = comm.rank(.pbd_env$SPMD.CT$comm))
}
\arguments{
\item{adios.rank}{}

\item{adios.rank}{}
}

