% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adios-package.r
\docType{package}
\name{pbdADIOS-package}
\alias{pbdADIOS-package}
\title{pbdADIOS-package}
\description{
Package to read from ADIOS, including staging.
}
\author{
Pragneshkumar Patel, George Ostrouchov, Wei-Chen Chen, Drew Schmidt
}
\keyword{package}

